Extracted cart parts to its own component. Created Your Cart for header for both mobile and desktop. Corrected review text color clash warning
