+++ src/components/header/Cart.tsx
@@ -1,6 +1,5 @@
 'use client';
 import { useCartContext } from '@/providers/CartProvider';
-import { Button } from '@/components/ui/button';
 import {
   Sheet,
   SheetClose,
@@ -10,72 +9,75 @@ import {
   SheetTitle,
   SheetTrigger,
 } from '@/components/ui/sheet';
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from '@/components/ui/card';
 import { TCartItems } from '@/lib/cart/useCart';
-import Image from 'next/image';
-import Link from 'next/link';
 import { HiOutlineShoppingCart } from 'react-icons/hi';
-import { FaCheck } from 'react-icons/fa6';
-import { IoIosClose } from 'react-icons/io';
+import { IoClose } from 'react-icons/io5';
+import AddToCartBody from '../cart/AddToCartBody';
+import AddToCartFooter from '../cart/AddToCartFooter';
+import YourCartHeader from '../cart/YourCartHeader';
+import { useMediaQuery } from '@mui/material';
+import BottomUpDrawer from '../ui/bottom-up-drawer';
 
 function Cart() {
   const { cartItems, cartOpen, setCartOpen } = useCartContext();
   const cartColor = cartItems.length > 0 ? '#BE1B1B' : '#000000';
+  const isMobile = useMediaQuery('(max-width: 768px)');
+
   return (
     <>
-      <Link href="/checkout" onClick={() => setCartOpen(false)}>
-        <HiOutlineShoppingCart
-          size={24}
-          color={cartColor}
-          className="mt-1 flex h-[40px] w-5 items-center hover:cursor-pointer md:order-last"
-          onClick={() => setCartOpen(!cartOpen)}
-        />
-      </Link>
-      <ItemsInCartAnimation cartItems={cartItems} />
-      {/* <Sheet open={cartOpen}>
-        <SheetTrigger asChild></SheetTrigger>
-
-        <SheetContent className="overflow-scroll">
-          <SheetHeader>
-            <SheetTitle className="flex w-full items-center justify-between">
-              <div className="text-green-600 *:h-5 *:w-5">
-                <FaCheck />
-              </div>
-              Your Cart
-              <SheetClose
-                asChild
-                className="cursor-pointer bg-gray-200 text-black *:h-8 *:w-8"
-              >
-                <button
-                  className="rounded-full"
-                  onClick={() => setCartOpen(false)}
+      <Sheet open={cartOpen}>
+        <SheetTrigger asChild>
+          <>
+            <HiOutlineShoppingCart
+              size={24}
+              color={cartColor}
+              className="mt-1 flex h-[40px] w-5 items-center hover:cursor-pointer md:order-last"
+              onClick={() => setCartOpen(!cartOpen)}
+            />
+          </>
+        </SheetTrigger>
+        <ItemsInCartAnimation cartItems={cartItems} />
+        {isMobile ? (
+          <BottomUpDrawer
+            title={<YourCartHeader />}
+            open={cartOpen}
+            setOpen={setCartOpen}
+            footer={<AddToCartFooter />}
+          >
+            <AddToCartBody />
+          </BottomUpDrawer>
+        ) : (
+          <SheetContent className="flex flex-col">
+            <SheetHeader>
+              <SheetTitle className="flex w-full items-center justify-between py-7 pl-4 pr-7">
+                <YourCartHeader />
+                <SheetClose
+                  asChild
+                  className="cursor-pointer bg-gray-200 text-black *:h-6 *:w-6"
                 >
-                  <IoIosClose />
-                </button>
-              </SheetClose>
-            </SheetTitle>
-          </SheetHeader>
-          {cartItems.map((item) => {
-            return <CartLineItem key={item.sku} item={item} />;
-          })}
-          <SheetFooter>
-            <SheetClose asChild>
-              <Button type="submit" className="mt-10 w-full" asChild>
-                <Link href="/checkout" onClick={() => setCartOpen(false)}>
-                  Checkout
-                </Link>
-              </Button>
-            </SheetClose>
-          </SheetFooter>
-        </SheetContent>
-      </Sheet> */}
+                  <button
+                    className="rounded-full"
+                    onClick={() => setCartOpen(false)}
+                  >
+                    <IoClose />
+                  </button>
+                </SheetClose>
+              </SheetTitle>
+              <div className="border-b bg-white shadow-[0_4px_4px_0px_rgba(0,0,0,0.1)]"></div>
+            </SheetHeader>
+            <div className="mx-auto flex h-screen w-full flex-col overflow-y-scroll px-4 pt-20">
+              <AddToCartBody />
+            </div>
+            <div className="w-full bg-white shadow-[0_-4px_4px_-0px_rgba(0,0,0,0.1)]">
+              <SheetFooter>
+                <SheetClose asChild>
+                  <AddToCartFooter />
+                </SheetClose>
+              </SheetFooter>
+            </div>
+          </SheetContent>
+        )}
+      </Sheet>
     </>
   );
 }
